+++ source/renderer/app/themes/daedalus/light-blue.js
@@ -115,10 +115,10 @@ export const LIGHT_BLUE_THEME_OUTPUT = {
       'rgba(250, 251, 252, .3)',
   },
   buttonFlat: {
-    '--theme-button-flat-background-color': '#f0f3f5',
-    '--theme-button-flat-background-color-hover': '#f3f5f7',
-    '--theme-button-flat-background-color-active': '#ebeef0',
-    '--theme-button-flat-background-color-disabled': 'rgba(240, 243, 245, 0.3)',
+    '--theme-button-flat-background-color': 'rgba(68, 91, 124, 0.1)',
+    '--theme-button-flat-background-color-hover': 'rgba(68, 91, 124, 0.05)',
+    '--theme-button-flat-background-color-active': 'rgba(68, 91, 124, 0.12)',
+    '--theme-button-flat-background-color-disabled': 'rgba(68, 91, 124, 0.1)',
     '--theme-button-flat-text-color-disabled': '#5e6066',
     '--theme-button-flat-text-color': '#5e6066',
     '--theme-button-flat-outline-color': '#e1e4e6',
@@ -127,10 +127,10 @@ export const LIGHT_BLUE_THEME_OUTPUT = {
     '--theme-button-primary-background-color': '#243e62',
     '--theme-button-primary-background-color-hover': '#2f496e',
     '--theme-button-primary-background-color-active': '#1e304a',
-    '--theme-button-primary-background-color-disabled': 'rgba(36, 62, 98, 0.3)',
+    '--theme-button-primary-background-color-disabled': '#243e62',
     '--theme-button-primary-text-color-disabled': '#fafbfc',
     '--theme-button-primary-text-color': '#fafbfc',
-    '--theme-button-primary-outline-color': '#4a7ec7',
+    '--theme-button-primary-outline-color': 'rgba(255, 255, 255, 0.3)',
   },
   connecting: {
     '--theme-connecting-background-color': '#1c3e65',
@@ -779,10 +779,6 @@ export const LIGHT_BLUE_THEME_OUTPUT = {
     '--theme-utxo-tooltip-text-color': '#fafbfc',
   },
   recoveryPhrase: {
-    '--theme-recovery-phrase-normal-background-color': 'transparent',
-    '--theme-recovery-phrase-normal-border-color': 'rgba(68, 91, 124, .07)',
-    '--theme-recovery-phrase-warning-background-color':
-      'rgba(68, 91, 124, 0.05)',
     '--theme-recovery-phrase-attention-background-color':
       'rgba(234, 76, 91, .1)',
   },