+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
 ### Fixes
 
 - Fixed green Cardano theme white color and borders color ([PR 1584](https://github.com/input-output-hk/daedalus/pull/1584))
+- Fixed flat button color  ([PR 1586](https://github.com/input-output-hk/daedalus/pull/1586))
 
 ## 0.15.0